Create global demo presenter that can be tested.
