+++ demo/src/main/java/com/novoda/demo/ControllerView.java
@@ -1,4 +1,4 @@
-package com.novoda.demo.controller;
+package com.novoda.demo;
 
 public interface ControllerView {
